+++ Tests/Code/Helper_Functions/simulateDynamics.m
@@ -1,7 +1,6 @@
-% Copyright 2018 - 2020, MIT Lincoln Laboratory
-% SPDX-License-Identifier: X11
-%%
 function results = simulateDynamics(sample)
+% Copyright 2018 - 2021, MIT Lincoln Laboratory
+% SPDX-License-Identifier: X11
 % This function simulates the input encounter (sample)'s dynamics using
 % run_dynamics_fast.c
 
@@ -13,7 +12,11 @@ ic2 = [0,sample.v_ftps(2),sample.n_ft(2),sample.e_ft(2),sample.h_ft(2),sample.he
 event1 = sample.updates(1).event;
 event2 = sample.updates(2).event;
 
+% Dynamic constraints
+dyn1 = [1.7 1116 -10000 10000 deg2rad(3), 1000000];
+dyn2 = [1.7 1116 -10000 10000 deg2rad(3), 1000000];
+
 % Simulate dynamics
-results = run_dynamics_fast(ic1,event1,ic2,event2,sample.runTime_s);
+results = run_dynamics_fast(ic1,event1,dyn1,ic2,event2,dyn2,sample.runTime_s);
 
 end